+++ tests/Unit/SymfonyCache/PurgeListenerTest.php
@@ -71,7 +71,7 @@ class PurgeListenerTest extends TestCase
 
         /** @var Psr6Store $store */
         $store = \Mockery::mock(Psr6Store::class)
-            ->shouldReceive('prune')
+            ->shouldReceive('clear')
             ->once()
             ->getMock();
         $kernel = $this->getKernelMock($store);
@@ -93,18 +93,15 @@ class PurgeListenerTest extends TestCase
         /** @var StoreInterface $store */
         $store = \Mockery::mock(StoreInterface::class);
         $kernel = $this->getKernelMock($store);
-
         $purgeListener = new PurgeListener();
         $request = Request::create('http://example.com/', 'PURGE');
         $request->headers->set('Clear-Cache', 'true');
         $event = new CacheEvent($kernel, $request);
-
         $purgeListener->handlePurge($event);
         $response = $event->getResponse();
-
         $this->assertInstanceOf(Response::class, $response);
         $this->assertSame(400, $response->getStatusCode());
-        $this->assertSame('Store must be an instance of Toflar\Psr6HttpCacheStore\Psr6StoreInterface. Please check your proxy configuration.', $response->getContent());
+        $this->assertSame('Store must be an instance of Toflar\Psr6HttpCacheStore\ClearableInterface. Please check your proxy configuration.', $response->getContent());
     }
 
     public function testPurgeAllowedMiss()